1.06.22 dodanie generatora aut

+++ src/com/company/Owner.java
@@ -2,7 +2,7 @@ package com.company;
 
 public class Owner {
     public String name;
-    public Integer cash = 150000;
+    public Double cash = 150000.0;
     public Car car;
 
     public Owner() {
